- still a WIP but I added UI classes and udpated the View Model and Presenter to Richard's earlier PR =========== Parts that still need work - tests - scrollable view for list of posts - ViewPostGateway
