+++ src/main/java/view_post/interface_adapters/View.java
@@ -1,4 +1,4 @@
-package viewpost.interface_adapters;
+package view_post.interface_adapters;
 
 public interface View {
     void display(ViewPostViewModel viewModel);
