+++ src/main/java/view_post/ui/MainPostView.java
@@ -1,15 +1,20 @@
-package viewpost.ui;
+package view_post.ui;
 
 import javax.swing.*;
 
 public class MainPostView {
-    private JFrame mainpostview;
+    private JPanel mainpostview;
     private ViewPostView viewPost;
     private PostListView postList;
+    private HeaderView header;
 
     public MainPostView(){
-        this.mainpostview = new JFrame();
+        this.mainpostview = new JPanel();
         this.mainpostview.setSize(900,800);
+        // adding header view
+        this.header = new HeaderView();
+        this.mainpostview.add(header.getHeader());
+        this.header.getHeader().setBounds(0, 0, 900, 120);
         // adding view post view
         this.viewPost = new ViewPostView();
         this.mainpostview.add(this.viewPost.getViewPostPanel());
