--- src/main/java/viewpost/ui/HeaderView.java
@@ -1,4 +1,0 @@
-package viewpost.ui;
-
-public class HeaderView {
-}
