--- src/main/java/viewpost/ui/PostListView.java
@@ -1,4 +1,0 @@
-package viewpost.ui;
-
-public class PostListView {
-}
