--- src/main/java/viewpost/use_case/ViewPostOutputBoundary.java
@@ -1,7 +1,0 @@
-package viewpost.use_case;
-
-import viewpost.interface_adapters.ViewPostViewModel;
-
-public interface ViewPostOutputBoundary {
-    void updateActivePost(ViewPostResponseModel responseModel);
-}
