--- src/main/java/viewpost/use_case/ViewPostResponseModel.java
@@ -1,4 +1,0 @@
-package viewpost.use_case;
-
-public class ViewPostResponseModel {
-}
